Update patch set 6
Patch Set 6:

So far this discovery ctrlr only be used from the nvme-cli and that's our previous working on not expose this controller.

In the other patch https://review.gerrithub.io/#/c/366877/ , it adds the ability to let the upper caller decide whether to attach it or not in the probe_cb.

On the other hand, in any case the upper caller needs to discover (or list) those actual subsystems, this discovery ctrlr will needs to be exposed to the upper caller. This is where nvme-cli discover command works.

My initial change on the other patch is to add some flag in the opts to let the upper caller to set a notification to expose this ctrlr and by default our existing callers will not expose it. After reviewing with Daniel, that change looks like not generic and somewhat complicated. So that we still reply on the probe_cb(). In this way, most of our existing callers always return "true" on this probed device. And this patch has a general check/helper function to hide the discovery ctrlr if not needed which is by default now. The already merged nvme-cli code will return true in this case.

Patch-set: 6
Label: Code-Review=0