blob: add check for supported underlying dev block lengths
Fail spdk_bs_init() if the dev being used has an
LBA size that is larger than a metadata page or not evenly
divisible by the size of a metadata page.

There's no UT coverage currently for the changed file so no UT
was added, manually tested with hardcoded invalid block lengths
and hello_blob. Will put this file on the list of UT TODO on
Trello.

Change-Id: I0e0ca747ecd5b6039c20fb6a885382bde4527158
Signed-off-by: Paul Luse <paul.e.luse@intel.com>