blobfs: Remove file->open_for_write
I do not think that open_for_write is a useful flag should
be used in cache allocation. Also if there are mixed read
and write issued by one thread, once there is read, we consider
we can select this file as a victim, the algorithm seems not
quite fair, so remove it.

Change-Id: Iedaecf577c2193f3349f71818efdbafa11a93796
Signed-off-by: Ziye Yang <ziye.yang@intel.com>