Update patch set 2
Abandoned

With the previous patch to always init the ctrlr from RDMA connection, this issue is not reproduced. The new NVMe-oF ctrlr will be reported back for this application.

Abandon this first and will create a new patch when finding the issues.

Thanks.

Patch-set: 2
Status: abandoned
Tag: autogenerated:gerrit:abandon