Update patch set 1
Patch Set 1:

I elected to enable this functionality via specifying a new optional cmd-line argument '-a' instead of performing this as the default behavior.  Why?  First, without the new cmd-line, perf behaves as it did previously so this shouldn't impact any existing tests.  Further, the default behavior may present the case of overrunning the default number of I/O requests depending on queue depth and io size thereby enabling a test of behavior when this scenario occurs.

Patch-set: 1