Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Hi Lance,

I'm leaning towards just picking a bigger number of requests by default, and not requiring the user to select "-a" to get this behavior.  That will make it easier to use out of the box for the vast majority of users.

There are also NVMe devices that get performance if the driver splits I/O not just based on max I/O size, but also device-specific stripe boundaries.  A lot the Intel NVMe SSDs for example get better performance if I/O are split on 128KB LBA boundaries - and the SPDK driver does this splitting.  So if we are going to try to account for the total max number of requests that may be needed, we need to account for that as well.  That is heavily dependent on I/O size though - for example, random 4KB, 8KB, 16KB, 32KB, 64KB, 128KB I/O will never need to get split because of this (because we pick LBAs that are aligned to those I/O sizes), but something like 12KB might need to get split.

-Jim

Patch-set: 1
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1