Update patch set 2
Patch Set 2: Code-Review-1

Hi Cunyin,

Please add unit tests to cover these cases.

Also what happens if user deletes a file with a handle open, then the application exits (i.e. crashes) unexpectedly before the last handle is closed?  The next time the blobstore is opened, this file will show up again.

We discussed this offline - you will need to set and sync an xattr denoting that the file is deleted when there are refs open.  This xattr needs to be checked when opening the blobstore, so that it can delete the file/blob if it crashed when last open.

-Jim

Patch-set: 2
Label: Code-Review=-1