nvmf_tgt: remove use of MAX_VIRTUAL_NAMESPACE
The configuration file parser now uses its own (arbitrary) limit on the
number of namespaces; this removes the need for MAX_VIRTUAL_NAMESPACE to
be in the public API.

The new MAX_NAMESPACES limit in conf.c is larger than the current
MAX_VIRTUAL_NAMESPACE; any namespaces beyond the current limit will
fail.  Later patches will allow dynamically increasing the number of
namespaces in a subsystem.

Change-Id: Ice508ac0fee34622cd738b8f477f2dc20870d5f0
Signed-off-by: Daniel Verkamp <daniel.verkamp@intel.com>