Update patch set 4
Patch Set 4: Code-Review-1

This is not the direction we want to go. If the blobstore loses power and isn't able to write out the bit masks, it can detect that when it starts up because the super block will be in the dirty state. In that case, the blobstore needs to read the entire metadata region, page by page, to reconstruct the bit masks. It's expensive to do that, but it only occurs after a power failure.

Patch-set: 4
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1