Update patch set 8
Patch Set 8: Code-Review-1

Only tagging this for now because I think some more explanation is needed in the commit messages.  We really shouldn't be submitting anything with nothing but a title unless it's correcting a doc or adding a straightforward UT case or something.  I will defiantly take some time to review this whole series though as I've recently been going through a lot of these code paths to get a better understanding of the whole thing.  thanks!!

Patch-set: 8
Label: Code-Review=-1