Update patch set 10
Patch Set 10:

So here's what doesn't make sense to me, thanks for adding the commit msg by the way:

currently we don't support coming up dirty so in the completion routine that you've modified the disk state is already clean, it has to be right?  After loading we're either clean or dirty and if clean what condition are you trying to solve by immediately turning around and sync'ing something that is already in sync. If we load and see that the state is dirty then we'd have to address that condition and *then* re-sync.  I haven't gone through this in complete detail yet so if I'm missing something please let me know.  Thanks!

Patch-set: 10
Label: Code-Review=0