Update patch set 11
Patch Set 11:

Hi Paul, you are right, we don't support coming up from last dirty shutdown(which means we do not call the spdk_bs_unload() last time) in this patch(I have added the dirty shutdown and recover support in the following patch), so the process now is:
if we unload the blobstore by calling spdk_fs_unload() explicitly, we will sync the "clean" flag as 1 to disk in function _spdk_bs_unload_write_used_clusters_cpl(), which means next time when we load the blobstore, we could check the flag in supper block, and we know last time the blobstore is clean shutdown, then we could just build all blobs in normal ways, but for dirty shutdown, we should set the clean flag to 0 (which means dirty shutdown) and sync the "clean" flag to disk as soon as we load the blobstore, because the system could crushed at anytime, we only modify the clean to 1 if we call the spdk_bs_unload(), otherwise, it means the blobstore is dirty shutdown. we need do the recover work.

Patch-set: 11
Label: Code-Review=0