Update patch set 16
Patch Set 16: Code-Review-1

Yeah, the comment "sync the clean flag to disk as soon as we load the blobstore, because the system could crushed at anytime, we only modify the clean to 1 if we call the spdk_bs_unload()" is what I am not in agreement with and consistent with what Ben just said. As soon as you load the BS you effectively switch over to the in memory structs and consider the state dirty until sync'd.  What you are proposing here is a bit of an optimization geared towards the case where the BS comes up dirty. I haven't looked at your next patch but I would think what we'd want to do is save the sync from dirty->clean for post recovery (not every load) in addition to what Ben state about on unload.  Ben you state "only on a call to unload" but I assume that because today there is no recovery/rebuild code but once that's in place would you agree w/sync at that point?

Patch-set: 16
Label: Code-Review=-1