Update patch set 14
Patch Set 14: Code-Review-1

(3 comments)

Couple of small things and I didn't go through all of the rebuild logic just yet since I think you need to rework things just a bit anyway based on the feedback on the patch this depends on.  Great job jumping on this though, note that it's on trello (as well as the crc thing) so you should go assign these to yourself and put them in the 'doing' column also.

Patch-set: 14
Label: Code-Review=-1