io_channel: defer unregister callback
In blobstore bd_dev is being used aftrer being freed.
Destroy of bs_dev has to be postoponed. This is generic
solution for any future implementations.

There already has been patch that fixed that issue, but it has
introduced new issue.
Scenario:
1.Blobstore on top of a malloc bdev
2.Attempt to remove malloc bdev
3.Malloc bdev unregister calls blobstore remove callback (hypothetical
  -not implemented yet)
4.Malloc bdev->status = SPDK_BDEV_STATUS_REMOVING
5.Blobstore remove calls spdk_bdev_close -> spdk_bdev_unregister on
  Malloc bdev
And now where the problem existis:
Blobstore remove calls _spdk_bs_free->spdk_io_device_unregister where
we lock the mutex on g_devlist_mutex and before we unlock it we call
unregister callback (point 5.) which goes to spdk_io_device_unregister
and tries to lock the same mutex again. Because of that we cannot call
_spdk_bs_dev_destroy as a callback to spdk_io_device_unregister – we
need to finish unregistering io device first so that we don’t stuck on
mutex lock.

Signed-off-by: Maciej Szwed <maciej.szwed@intel.com>
Change-Id: Ib6c290e3d6d05febdb6027e1db064a92c6926b41