Update patch set 18
Patch Set 18: Code-Review-1

So the CRC field is not implemented right now, it just exists. I don't believe adding a flag here is the right approach but will defer to one of the maintainers.  Same thing for test code, I believe independent test code make a lot of sense here (not code that relies on blobfs, that's a good test case for the blobfs module) so I'd prefer to see UT added to this as well.

Patch-set: 18
Label: Code-Review=-1