Update patch set 17
Patch Set 17: Code-Review-1

interesting, I was certain at one point during creating the CLI that I ran into a dirty scenario on load.  On quick review I'm not sure how that could have happened.  I need to look at this a little closer in that regard but I still do not suspect what you are doing here is the best approach.  More later, today is a US holiday so it may be tomorrow before I get back to this.

Patch-set: 17