Update patch set 17
Patch Set 17:

So I haven't thought this all the way through but what's sorta bugging me - is the clean polarity backwards here? IOW, seems like we should mark the on disk struct dirty on creation and on load and then update to clean when sync'd, unloaded or following a rebuild. Currently, well I may be missing something, when is it ever marked dirty? I'm pretty sure this is what did did back in the RAID days and had a timer to go clean every 4 seconds or whatever but that was the idea, on disk is dirty when the project is alive and only clean when made so purposefully

Patch-set: 17