Update patch set 17
Patch Set 17:

Hi Paul, "seems like we should mark the on disk struct dirty on creation and on load and then update to clean when sync'd", that is exactly, so this is the patch try to do. We only do sync the clean flag when we unload the blobstore now.

Patch-set: 17