Update patch set 17
Patch Set 17:

Hi Paul, for your comments:
or maybe it makes more sense to mark it clean on disk when:
1) it's sync'd
   this is too expensive, let's take p4800 as example and format the block as 4K, it will use 22 pages for bit masks, we should sync the "clean" flag after we sync the used bit mask, but I think we do not want to sync the used md and used cluster bit mask for every md sync operation in run time.
2) clean shutdown
we have did this work now, in spdk_bs_unload().
3) after recovery
after recovery, we are in run time, the app could be crashed anytime, we need sync the the blobstore as dirty, not clean.
and mark it dirty when:
1) at the start of any update to volatile metadata structs
That means we should read the supper block every time when we try to sync md? it is expensive.

Patch-set: 17