Update patch set 6
Patch Set 6:

> (1 comment)

I think that it is possible to use package.

But we need to change pmem library to get better performance, so for until we merge our patch I think it will be easier to use submodule.

The only thing that I will change is that nvml submodule will be disabled by default (until user will specify ./configure --with-nvml)

Patch-set: 6
Label: Code-Review=0