Update patch set 17
Patch Set 17: Code-Review+2

This patch is right. There was an oversight in the code where it marked the in memory representation dirty on load, but never wrote that to disk. This fixes that. It works the way Paul outlined - toggles dirty on load and clean on unload, but I had forgotten to write the toggle to disk on load.

Patch-set: 17