Update patch set 18
Patch Set 18:

(2 comments)

Agree with Paul.

We should instead implement the CRC on the page and save that to SSD.  Let's do that as a first patch.

Determining which pages are valid on dirty shutdown is actually much more complicated than what is implemented here.  We need to consider the case of blobs which require multiple pages of metadata.  We could have a blob with md page 1 that points to md page 2 for example.  If this blob's metadata gets updated, we need atomic updates, so we cannot update both page 1 and page 2 simultaneously.  Instead we allocate a new md page 3, write that to disk then once that completes, we update page 1 to point to page 3.  If we crash before this final write, page 3 will appear valid but nothing points to it so it should *not* be considered used.

Patch-set: 18