Update patch set 9
Patch Set 9:

Hi Daniel,

Thank you for your comments.
Regarding the blocking issue, we'll take a look at making the update asynchronous.
As for your other 2 comments, we tried to address this by opening all bdevs associated with the controller before attempting the update. If any bdevs for the controller are already open, we disallow the update. Please let us know if what we have is sufficient.

Patch-set: 9
Label: Code-Review=0