Update patch set 10
Patch Set 10:

This is how I wanted to do this, but I had concern about thread that we use from __get_thread() function. It takes current thread, but thread used in spdk_bs_unregister_md_thread() function (which was the cause why we needed to defer destruction) is not the current thread, but the thread that was assigned to channel that we destroying. I was worried that there might be some race conditions when we use 2 different threads, but I noticed that we actually use current thread as a channel thread also so the problem does not exists.

Patch-set: 10
Label: Code-Review=0