Update patch set 7
Patch Set 7:

I'm not sure where the submodule path is coming from (I see that the .gitmodules file is now pointing at nvml on github). This may be a build_pool bug.

However, when I clone this patch myself, it also fails, because the submodule is pointing at a commit that doesn't exist in the upstream pmem/nvml.git repository.

Patch-set: 7