Update patch set 10
Patch Set 10:

I think the usual strategy is to try READ CAPACITY (10) first and only issue READ CAPACITY (16) if it returns 0xFFFFFFFF - this is so that older devices that only support READ CAPACITY (10) don't get confused by the newer command and fail in some way, so it's probably not really relevant with virtio (the target probably always understands READ CAPACITY (16)).  Nevertheless, it might be a good idea to stick with the usual 10 -> 16 pattern to avoid potential problems.

Patch-set: 10
Label: Code-Review=0