Update patch set 4
Patch Set 4: Code-Review-1

(3 comments)

It would be good to add autodetection support for this to scripts/autotest_common.sh so we can turn on the configure flag when the libraries are installed in the automated builds.  Look for the section that sets config_params based on the existence of e.g. /usr/include/infiniband/verbs.h

Patch-set: 4
Label: Code-Review=-1