Update patch set 5
Patch Set 5:

"It would be good to add autodetection support for this to scripts/autotest_common.sh so we can turn on the configure flag when the libraries are installed in the automated builds.  Look for the section that sets config_params based on the existence of e.g. /usr/include/infiniband/verbs.h"
This is done in next patch in series -> https://review.gerrithub.io/#/c/379434/

Patch-set: 5
Reviewer: Tomasz Zawadzki <1011221@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0