make: changes to support dpdk submodule with a relative path
Make the dpdk submodule in spdk the default dpdk library. The dpdk submodule
is located in the dpdk director at the top level of the spdk tree.

Following this "make", "make clean" and "make CC=clang CXX=clang++" will automatically
build, clean and rebuild dpdk at the same time that it builds spdk. This leverages
the integrated support for dpdk in the spdk Makefiles.

To overide this behavior to use the old method, simply say:

 make CONFIG_DPDK_DIR=../dpdk/build

This will ignore the dpdk submodule in spdk, even it's it there.

Change-Id: Id2db0ee4192967de1d66708be94a4de4d33ef506
Signed-off-by: John Meneghini <johnm@netapp.com>