Update patch set 8
Patch Set 8:

> (10 comments)
 > 
 > way cool Seth! Some little stuff throughout, a few general
 > questions though:
 > 1) did you consider just modifying hello_blob with a compilation
 > option to use either bdev+event framework or direct nvme? There's a
 > lot of common code and the different stuff is pretty isolated so it
 > could come out pretty clean. Nothing wrong with another separate
 > example though but but arguably with this much in common it could
 > cause confusion too. Just a thought
 > 2) did you talk to Ben about making the blob_nvme.* files part of
 > the lib instead of part of the example? I think there'd be some
 > benefit there. If they are going to stay here I'd put all the cpde
 > in one file and clearly identify the section that is the that "back
 > end" portion of BS for nvme. These filenames are very confusing
 > 3) Either way you don't need a separate blob_nvme.h file, you can
 > put those few things in the C file as there not shared with
 > anything else.
 > 4) I didn't check all the error paths, you're missing a few checks
 > on allocation failures, but for hello_blob I put brkpnts in every
 > place there could be a failure and 'inserted one' so that I could
 > be sure they all worked as expected and found a few issues towards
 > the end. As this is example code please make sure you test every
 > path, we don't want someone basing their code of something they
 > copied and pasted here that wasn't correct and without test code
 > it's all on you to make sure it's right.
 > 
 > Also, is this done yet? Great job man! Someone was asking about
 > direct nvme with BS not too long ago, great to have an example now

Hi Paul, Thanks for the great advice! I have moved the nvme_blob.c code to lib because I think it will be useful for customers.
I think that using the same example with different compile flags is a cool option (because admittedly almost the whole callback chain is from your example), I honestly didn't think of doing it that way, and my only holdback would be that it may be easier for a customer trying to use the examples as a base to see the two methods in isolation. 
Thanks for the heads up about the error paths. I fixed some of them in this revision, and will do another sweep to make sure that I catch all of them.

Patch-set: 8
Label: Code-Review=0