[WIP]Add py-spdk client for SPDK
As mentioned before, SPDK is a high-performance kit written in c.
It is hard for management-level applications written in python to
connect with SPDK-based app (as: nvmf_tgt, vhost, etc.) directly.
For this reason, developers should be able to provide a client for
callers to make convenient use of SPDK-based app. The py-spdk is
such a python client that is designed for the upper-level
management applications to communicate more fully and get data
returned by SPDK-based app.

In this situation, we can explain how py-spdk works by the example
of OpenStack Cyborg [0]. The py-spdk can make the Cyborg interact
with SPDK-based app. The pyspdk.py uses the rpc.py to communicate
with the nvmf_tgt, and then to get something returned (as:
get_luns, get_interfaces, get_vhost_blk_controller, etc.)[1].

[0] https://review.openstack.org/#/c/498690/
[1] https://github.com/hellowaywewe/py-spdk/blob/master/README.md

Change-Id: I0bc12fa9618d75b4b8c5cfddcbc2e722b58541a6
Signed-off-by: heluwei <simple_hlw@163.com>