[WIP]Add py-spdk client for SPDK
As mentioned before, SPDK is a high-performance kit written in c.
It is hard for management apps written in python to connect with
SPDK-based appdirectly. For this reason, developers should be able
to provide a client for management apps to make convenient use of
SPDK-based app. The py-spdk is a client that is designed for the
management apps which built upon SPDK. It provides a general python
binding for spdk framework which means SPDK app developers could
quickly develop a SDK based upon py-spdk.

In this situation, we can take OpenStack Cyborg as a use case to
explain how py-spdk works [0]. The py-spdk can make the Cyborg interact
with SPDK-based app. In the py-spdk, the pyspdk.py will communicate with
the SPDK-app through the rpc.py. It provides some fundamental functions
including the judgment of process status, hugepages initialization and
server connection [1].

The py-spdk has been implemented two kinds of client to do some
operations (such as: get_luns, get_interfaces, etc) to SPDK-based app
which are nvmf_client and vhost_client respectively [1]. The third
SPDK-based app (iscsi_client) will be added later.

Notice: The py-spdk is currently in favour of SPDK v17.07 and DPDK v17.08.

[0] https://review.openstack.org/#/c/498690/
[1] https://github.com/hellowaywewe/py-spdk/blob/master/README.md

Change-Id: I0bc12fa9618d75b4b8c5cfddcbc2e722b58541a6
Signed-off-by: heluwei <simple_hlw@163.com>