[WIP]Add Blobkv.c to support kv
The blobkv.c in which serveral stub functions have been
written is submitted. It is not rigorous and still being
modified.

For the blobkv[0], we should be able to add a hashtable into
the memory to perform matching with key.

When blobkv receives the request of GET, firstly compare the
k of function kv_blob_read() with the key of the hashtable.
If the match is successful, it returns the blob’s address.
And then call the function in the blobstore to read data from
the specific blob.

When blobkv receives the request of PUT, firstly validate the
size of v in the function kv_blob_write(). If the data is
oversize, call the function kv_blob_write_part() to execute
partial storage. Before storing data v, write the k into the
hashtable, and then call the function in the blobstore to
write data on the blob.

When blobkv receives the request of DELETE, like PUT.

To make blobstore not to do changes as much as possible
to support object storage, we introduced the blobkv_context
structure, as follows:
struct blobkv_context {
    struct spdk_blob_store *bs;
    struct spdk_blob *blob;
    spdk_blob_id blobid;
    struct spdk_io_channel *channel;
    uint8_t *read_buff;
    uint8_t *write_buff;
    uint64_t page_size;
};

[0] https://trello.com/b/P5xBO7UR/things-to-do

Change-Id: Iea361e5fddbe8af44ca8450b8333f8107e835a55
Signed-off-by: heluwei <simple_hlw@163.com>