Update patch set 3
Patch Set 3:

cool... I hate that we don't have UT for this or the other helper API for bdevs. In light of that maybe hold off on merging this until the example that uses it is fully ready to merge and tested?

Patch-set: 3
Reviewer: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=0