examples/perf: report an error when doing large I/O or high queue depth tests
For perf tests with large I/O size, such as 1MiB, even with queue
depth 64, the NVMe driver may split one request to several based
on stripe and maximum transfer size, so here we add a simple
caculation for the scenario which will report an error to users.
Also add a new parameter for the real queue pair size instead of
the queue depth which user specified for tests.

Change-Id: I66922a3d673dc97796d7fbe6e86cf5037a45b37d
Signed-off-by: Changpeng Liu <changpeng.liu@intel.com>