Update patch set 1
Patch Set 1: Code-Review-1

(5 comments)

I need to review this patch some more.  So far it is just some cosmetic changes.

It would be great to be able to break this up into smaller pieces to help with the review, but it is not clear to me yet how to split it up.

If you can think of a way to break this up into smaller patches that would be appreciated.  If not, I will come back and do a more thorough review on it.

Thanks,

-Jim

Patch-set: 1
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1