Update patch set 1
Patch Set 1:

Paul: I thought that the ref counting (that this patch is putting back) was only used to control when to free the I/O channel, which already has its own ref counting, but it appears that it was necessary for managing the lifetime of the LUN itself.

We should definitely add a test that can reproduce the crash reported in the GitHub bug so that we can prevent this from regressing.

Patch-set: 1
Label: Code-Review=0