Update patch set 1
Patch Set 1:

The other advantage of doing the uuid checking during load is that it will avoid loading the blobstore successfully, checking the uuid, and then doing an unload when it doesn't match.  Loading the blobstore successfully writes a dirty bit to catch unexpected power failures which require rebuilding the md and cluster bit masks.  But if we check the uuid immediately after reading the super block, we can avoid doing any writes when they aren't needed.

Patch-set: 1